internal, log: remove code for old unsupported go-versions
