+++ internal/debug/loudpanic.go
@@ -14,9 +14,6 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-//go:build go1.6
-// +build go1.6
-
 package debug
 
 import "runtime/debug"
